+++ packages/lib-react-native-jsi/cpp/tonclient.h
@@ -1,4 +1,4 @@
-// https://github.com/tonlabs/ton-client-js/blob/master/packages/lib-react-native/ios/tonclient.h
+// https://github.com/tonlabs/ever-sdk-js/blob/master/packages/lib-react-native/ios/tonclient.h
 
 #pragma once