+++ packages/lib-react-native/android/src/main/java/tonlabs/tonclient/TonClientJsonInterface.java
@@ -2,7 +2,7 @@ package tonlabs.tonclient;
 
 public class TonClientJsonInterface {
     static {
-        System.loadLibrary("tonclient");
+        System.loadLibrary("eversdk");
     }
 
     public interface IResponseHandler {