+++ packages/lib-web/lib/src/lib.rs
@@ -43,7 +43,7 @@ pub fn core_destroy_context(context: u32) {
     destroy_context(context)
 }
 
-#[wasm_bindgen(js_namespace = tonclient)]
+#[wasm_bindgen(js_namespace = eversdk)]
 extern "C" {
     fn core_response_handler(request_id: u32, params: JsValue, response_type: u32, finished: bool);
 }