+++ packages/tests-react-native/install-using-pack.js
@@ -35,7 +35,7 @@ function pack(names) {
 function copyBinary(relPath) {
     const srcPath = path.resolve(__dirname, '..', 'lib-react-native', ...relPath);
     if (fs.existsSync(srcPath)) {
-        const dstPath = path.resolve(__dirname, 'node_modules', '@tonclient', 'lib-react-native', ...relPath);
+        const dstPath = path.resolve(__dirname, 'node_modules', '@eversdk', 'lib-react-native', ...relPath);
         fs.mkdirSync(path.dirname(dstPath), { recursive: true })
         fs.copyFileSync(srcPath, dstPath);
         process.stdout.write(`Copy ${srcPath}\n`);
@@ -55,9 +55,9 @@ try {
         fs.unlinkSync(path.resolve(__dirname, `${tar}`));
     }
     for (const arch of ['arm64-v8a', 'armeabi-v7a', 'x86', 'x86_64']) {
-        copyBinary(['android', 'src', 'main', 'jniLibs', arch, 'libtonclient.so']);
+        copyBinary(['android', 'src', 'main', 'jniLibs', arch, 'libeversdk.so']);
     }
-    copyBinary(['ios', 'libtonclient.a']);
+    copyBinary(['ios', 'libeversdk.a']);
 } catch (error) {
     console.error(error.message);
 } finally {