+++ packages/tests-web/webpack.config.js
@@ -19,7 +19,7 @@ module.exports = {
     plugins: [
         new CopyWebpackPlugin({
             patterns: [
-                { from: './node_modules/@tonclient/lib-web/tonclient.wasm' },
+                { from: './node_modules/@eversdk/lib-web/eversdk.wasm' },
             ],
         }),
         new webpack.LoaderOptionsPlugin({
@@ -34,11 +34,11 @@ module.exports = {
                 include: [
                     path.resolve('index.js'),
                     path.resolve(__dirname, './node_modules/webpack-dev-server'),
-                    path.resolve(__dirname, './node_modules/@tonclient/core'),
+                    path.resolve(__dirname, './node_modules/@eversdk/core'),
                     path.resolve(__dirname, './node_modules/assert'),
                     path.resolve(__dirname, './node_modules/buffer'),
-                    path.resolve(__dirname, './node_modules/@tonclient/tests'),
-                    path.resolve(__dirname, './node_modules/@tonclient/lib-web'),
+                    path.resolve(__dirname, './node_modules/@eversdk/tests'),
+                    path.resolve(__dirname, './node_modules/@eversdk/lib-web'),
                 ],
             },
             {